requirements.txt
opencv-python==4.5.5.64
face_recognition
numpy==1.21.2
pillow==8.3.2
PyQt5==5.15.4
paho-mqtt==1.6.1
